backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
aiomysql==0.2.0
SQLAlchemy==2.0.35
